gen new vids

+++ gen.py
@@ -3,6 +3,7 @@ import numpy as np
 import os, subprocess, glob
 from matplotlib import pyplot as plt
 from scipy import ndimage
+import time
 
 import matplotlib.cm as cm
 
@@ -14,7 +15,7 @@ class Img3D:
         self.k = k
         self.img_unfiltered, self.centers = self.gen_3d_points()
         # Might need to change sigma based on n.
-        self.img_filtered = ndimage.gaussian_filter(self.img_unfiltered, sigma=3, order=0)
+        self.img_filtered = ndimage.gaussian_filter(self.img_unfiltered, sigma=7, order=0)
 
     def gen_3d_points(self):
         """
@@ -28,8 +29,8 @@ class Img3D:
         for _ in range(self.k):
             a = np.random.randint(self.n // 4, 3 * self.n // 4)
             b = np.random.randint(self.n // 4, 3 * self.n // 4)
-            # c = np.random.randint(self.n//4, 3*self.n//4)
-            c = self.n // 2
+            c = np.random.randint(self.n//4, 3*self.n//4)
+            # c = self.n // 2
             centers += [(a, b, c)]
             r1 = np.random.randint(self.r // 2, self.r)
             z, y, x = np.ogrid[-a:self.n - a, -b:self.n - b, -c:self.n - c]
@@ -39,36 +40,61 @@ class Img3D:
 
     def show_img(self, filter_flag):
         if filter_flag == 0:
-            plt.imshow(self.img_unfiltered[self.centers[0][0]], cmap='gray', interpolation='nearest');
+            plt.imshow(self.img_unfiltered[self.centers[0][0]], cmap='gray', interpolation='nearest')
         elif filter_flag == 1:
-            plt.imshow(self.img_filtered[:, self.centers[0][0], :, :], cmap='gray', interpolation='nearest');
+            plt.imshow(self.img_filtered[:, self.centers[0][0], :, :], cmap='gray', interpolation='nearest')
 
 
-def generate_video(img):
+def generate_video(obj, vid=0, filtered=True, plane="xy"):
     # clearing old files
     folder = 'output'
+    """
     for file in os.listdir(folder):
         file_path = os.path.join(folder, file)
         try:
-            if os.path.isfile(file_path):
+            if file[:5] != "video" and os.path.isfile(file_path):
                 os.unlink(file_path)
         except Exception as e:
             print(e)
+    """
+    if filtered:
+        img = obj.img_filtered
+    else:
+        img = obj.img_unfiltered
 
     # generating new pngs, mp4
     for i in range(len(img)):
-        plt.imshow(img[i], cmap=cm.Greys_r)
+        if plane == "xy":
+            plt.imshow(img[i], cmap=cm.Greys_r)
+        elif plane == "xz":
+            plt.imshow(img[:,i,:], cmap=cm.Greys_r)
+        elif plane == "yz":
+            plt.imshow(img[:,:,i], cmap=cm.Greys_r)
         plt.savefig('output/file%02d.png' % i)
     os.chdir('output')
     subprocess.call([
         'ffmpeg', '-framerate', '8', '-i', 'file%02d.png', '-r', '30', '-pix_fmt', 'yuv420p',
-        'video.mp4'
+        'video%d.mp4' % vid
     ])
 
     # i think this is supposed to remove files but it does not.
     for file_name in glob.glob("*.png"):
         os.remove(file_name)
+    os.chdir('..')
 
 
-img = Img3D(50, 8, 1)
-generate_video(img.img_unfiltered)
+img_test = Img3D(100, 8, 6)
+
+
+start = time.time()
+
+generate_video(img_test, 0, False, "xy")
+generate_video(img_test, 1, False, "xz")
+generate_video(img_test, 2, False, "yz")
+
+generate_video(img_test, 3, True, "xy")
+generate_video(img_test, 4, True, "xz")
+generate_video(img_test, 5, True, "yz")
+
+end = time.time()
+print(end - start)
